requirements.txt
requests
bs4
readability-lxml
pydantic
fastapi
python-dotenv